move test to new loading module
